+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/dao/VideoAidDao.java
@@ -6,8 +6,8 @@ public interface VideoAidDao {
 
     Long queryVideoAidMaxId();
 
-    Integer queryVideoAidById(Long id);
+    Long queryVideoAidById(Long id);
 
-    List<Integer> queryVideoAidsByIds(List<Long> ids);
+    List<Long> queryVideoAidsByIds(List<Long> ids);
 
 }
