+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/dao/VideoExDao.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface VideoExDao {
 
-    VideoEx queryVideoByAid(Integer aid);
+    VideoEx queryVideoByAid(Long aid);
 
     List<VideoEx> queryVideos(Integer vc, Integer start_ts, Integer end_ts, Integer activity, Integer recent,
                               String title, String up, String order_by, Integer desc,
