+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/service/VideoRecordService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface VideoRecordService {
 
-    List<VideoRecord> queryVideoRecords(Integer aid, Integer last_count, Integer start_ts, Integer end_ts,
+    List<VideoRecord> queryVideoRecords(Long aid, Integer last_count, Integer start_ts, Integer end_ts,
                                         Boolean limit, Integer pn, Integer ps);
 
-    Integer queryVideoRecordsCount(Integer aid, Integer start_ts, Integer end_ts);
+    Integer queryVideoRecordsCount(Long aid, Integer start_ts, Integer end_ts);
 
 }
