+++ v2_spring_backend/tdd/src/main/java/com/bunnyxt/tdd/service/user/UserLikeVideoService.java
@@ -5,11 +5,11 @@ import com.bunnyxt.tdd.model.user.UserLikeVideo;
 
 public interface UserLikeVideoService {
 
-    TddCommonResponse postUserLikeVideo(Long userid, Integer aid);
+    TddCommonResponse postUserLikeVideo(Long userid, Long aid);
 
-    TddCommonResponse deleteUserLikeVideo(Long userid, Integer aid);
+    TddCommonResponse deleteUserLikeVideo(Long userid, Long aid);
 
-    UserLikeVideo queryUserLikeVideo(Long userid, Integer aid);
+    UserLikeVideo queryUserLikeVideo(Long userid, Long aid);
 
-    Integer queryUserLikeVideoCount(Integer aid);
+    Integer queryUserLikeVideoCount(Long aid);
 }
